refactor: Remove middleware.go and move into main.go
